added BaseAlg class
